app/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
jinja2>=3.1.0
psycopg
numpy
beautifulsoup4
pypdf
httpx
lxml
sentence-transformers
huggingface_hub
transformers
psycopg
python-dotenv
PyPDF2
pdfplumber
scispacy
https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_core_sci_lg-0.5.4.tar.gz